+++ R/spark_shell.R
@@ -221,6 +221,12 @@ start_shell <- function(master,
       shell_args <- c(shell_args, "--packages", paste(shQuote(packages, type = shQuoteType), collapse=","))
     }
 
+    # add environment parameters to arguments
+    shell_env_args <- Sys.getenv("sparklyr.shell.args")
+    if (nchar(shell_env_args) > 0) {
+      shell_args <- c(shell_args, strsplit(shell_env_args, " ")[[1]])
+    }
+
     # add app_jar to args
     shell_args <- c(shell_args, app_jar)